feat(admin/products): index, show, edit products

+++ database/migrations/2022_01_03_100913_create_products_table.php
@@ -12,7 +12,7 @@ class CreateProductsTable extends Migration
             $table->id();
             $table->string('name');
             $table->text('description');
-            $table->float('price');
+            $table->bigInteger('price');
             $table->bigInteger('stock')->nullable();
             $table->foreignId('category_id')
                 ->nullable()
